ISSUE_TEMPLATE: add a requirement for WIP docs
During the 1.3 retrospective one of the things we identified was that UX
and docs were coming long after a feature was perceived to be "done".
Further, the proposals might have changed while the feature was being
built. A few examples of this are PetSet or RBAC which landed in 1.3 but
the user experience rough sketch would have been helpful to get everyone
understanding the feature and as a starting off point for writing "real
docs".