README: cleanups to make it more compact
We had some duplicated information in the README and some informatin
that really wasn't adding much value.

The number one question we have gotten is still "should I file a
feature". The number two point of contention on this repo is "what is
the process I should follow before filing a feature".

Make these two things the top points of the README and drop other things
that aren't really connected. In particular the coding section and
proposal sections weren't adding much value. The process for proposals
are terribly hard to discover but we shouldn't try to solve this here.