Port Troubleshooting Running Pods proposal to KEP
This is a port of the existing Ephemeral Containers proposal to a KEP
candidate. I've refocused the document on the API and kubelet changes,
and put changes to kubectl out of scope for this document. I've also
updated the API for changes we've made during API review.

This KEP draft is still marked provisional.

Original document: https://github.com/kubernetes/community/blob/master/contributors/design-proposals/node/troubleshoot-running-pods.md