Update terminology in Scheduling Framework
This clarifies and tweaks a few terms in the scheduling framework so we
can be more explicit about concurrency in the framework. Specifically,
the term "scheduling cycle" is shortened to refer to only the parts of a
"scheduling context" which are executed by the main thread. The term
"binding cycle" is introduced to refer to the portions which are always
run concurrently.