Modify GetTopologyHints() for per-resource hints in the TopologyManager
At present, there is no way for a hint provider to return distinct hints
for different resource types via a call to `GetTopologyHints()`. This
means that hint providers that govern multiple resource types (e.g. the
devicemanager) must do some sort of "pre-merge" on the hints it
generates for each resource type before passing them back to the
`TopologyManager`.

This seems counter-intuitive, since there is no practical reason that a
"pre-merge" should be necessary -- it just happens to be necessary
because of the way the current interface is designed.

It would be better to allow a hint provider to pass back raw hints for
each resource type, and allow the `TopologyManager` to merge them using
a single unified strategy.

This patch makes a simple change to the `GetTopologyHints()` interface
to allow this to occur.

Moreover, this change allows the `TopologyManager` to recognize which
resource type a set of hints originated from, should this information
become useful in the future.