Add Merge method to Topology Policy
In the current design the Topology Manager generates the merged
hints and select the best hint. The policy indicates whether or not
to admit the pod.

This proposal adds Merge abstraction so that the policy generates
the merged hints and the Topology Manager selects the best hint.
This gives the flexibility to add other polices such as align all
resources to one NUMA.

Signed-off-by: Moshe Levi <moshele@mellanox.com>