Fix hack/update-toc.sh
We mostly switched to go.mod, but hack/update-toc.sh was not updated.

We now don't really need to install the binary, we can simply `go run`
it (it can be slower, but for the one-off use case it shouldn't much
matter).

Rather than having each script `go get` at a particular SHA, it seems
easier just to use go.mod / go.sum.