initial generic ephemeral inline volumes KEP
This KEP proposes an alternative mechanism for specifying volumes
inside a pod spec. Those inline volumes then get converted to normal
volume claim objects for provisioning, so storage drivers do not need
to be modified.

The lifecycle of those volumes is ephemeral: they will get deleted
once the pod terminates.

Because this is expected to be slower than CSI ephemeral inline
volumes, both approaches will need to be supported.