Add numaid and cpus into PodResources interface
This change necessary for resource with topology exporting daemon,
which used in topology aware scheduling.

Information about CPU is keeping in cpu_ids, since it's  enough to
represent both quantity and numaid. NUMAid can be obtained from
cadvisor MachineInfo, since id in cpus_ids is a thread_id.
This API doesn't provide cpu fraction, since it could be obtainded from
Pod's request/limits and in case of non-integer CPU quantity and
non-guaranteed QoS cpu assigned is not exclusive and NUMA id is not
interesting.

Device assignment on the NUMA node is representing by Topology
structure. The structure was chosen because of several reason:
 more extensibility (with keeping backward compatibility)
 more compact representation of case when we have device aligned on
several NUMA nodes (compared to just numa_id field in ContainerDevices)

Signed-off-by: Alexey Perevalov <alexey.perevalov@huawei.com>