podresources: watch: expose resourceVersion
Add new field in the API responses objects to allow client applications
to consume both `List` and `Watch` endpoints.

The issue here is enabling client applications to not lose any updates
when both APIs are used.

The straightforward option is to follow the generic k8s approach (see
link below) and let kubelet keep a historical window of
the last recent changes, so client applications have the chance to
issue `List` and shortly after `Watch`, starting from the
resourceVersion returned in `List`.
The underlying assumption is indeed that `Watch` happens "shortly" after `List`,
otherwise the system cannot guarantee the lack of gaps.

However implementing this support requires to keep the aforementioned
sliding window of changes, which however requires careful implementation
to address scalability and safety guarantees.

However, the `podresources` API is a specific API, so, while is good to
follow as much as possible the generic API concepts, it also allows some
possible little differences which can help keep the implementation
simple and safe.

This patch proposes a simplest possible approach to reconcile the
`List` and `Watch` responses, providing the `resource_version` field and
suggesting a little change in the client applications programming model.

Inspired by the concepts found on
https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes

Signed-off-by: Francesco Romani <fromani@redhat.com>