Service Internal Traffic Policy: use more generic API
The current choice of `Cluster` will likely cause some confusion for implementations providing multicluster Service routing.

For example, Cilium and Istio both transparently route to all endpoints across all clusters (see https://cilium.io/blog/2019/03/12/clustermesh#service-discovery). This makes the statement that `Route to all cluster-wide endpoints (or use topology aware subsetting if enabled)` no longer correct for these implementations.

As a counter example, I believe Linkerd's model technically is compatible, as it is actually copying Endpoints over into the cluster. However, it may still confuse users as despite the Endpoint being in the cluster, the destination of the Endpoint is not, and the fact the Endpoint is in the cluster is an implementation detail.

Another counter example is MCSD, where the Service is still "cluster", and a separate ServiceExport is used to define the supercluster.

It is also possible this could instead be extended to have 3 options: Supercluster, Cluster, and Local, effectively replacing ServiceExport.

So this PR changes the naming a bit to allow for both models to exist gracefully