Update CRI KEP with pinned Image requirement
This reflects the new requirement of CRI going to Beta that the CRI implementation
must be able to indicate to the Kubelet an image should not be garbage collected.

This functionality will be used in leu of the Kubelet's current behavior
of keeping track of the pause image version to stop the pause image from being garbage collected.

Signed-off-by: Peter Hunt <pehunt@redhat.com>