klog flag deprecation: clarify scope and timeline
Deprecating the flags only in some binaries is not going far enough because the
full feature set would still have to be supported for the remaining ones. The
goal was to remove them everywhere and the KEP gets updated to reflect that.

Because of the deprecation period for command line flags we want to announce
the intent to deprecate as soon as possible. The proposal is to do it already
at the alpha stage in Kubernetes 1.23. Most of the work is done or can be
completed.

Depending on the interpretation of "must function for 2 releases after their
announced deprecation" (does that include the release for which the announcement
was made?) we might be able to remove in 1.25. But to be safe and also to
ensure the "for 12 months" rule, we aim for removal in 1.26.