KEP-3157: allow informers for getting a stream of data instead of chunking.
The kube-apiserver is vulnerable to memory explosion.
The issue is apparent in larger clusters, where only a few LIST requests might cause serious disruption.
Uncontrolled and unbounded memory consumption of the servers does not only affect clusters that operate in an
HA mode but also other programs that share the same machine.
In this KEP we propose a potential solution to this issue.