contextual logging: simplify scope, avoid transition
A less ambitious goal is to just remove the usage of the traditional klog
logging API and the hard dependency on the klog logging implementation and
still keeping k8s.io/klog/v2 as dependency.

The main advantage is that there's no need for a transition to a new
k8s.io/klogr package.