dynamic resource allocation: update status
The KEP discussion showed that a longer period in alpha seems likely, so now
the (of course still tentative) plan is to keep it in alpha for three releases.

SIG Autoscaling will need to be involved, now as approver and later as
reviewer, so they get added as "participating SIG".

Finally, the goal is to start working on this in Kubernetes 1.25, so it has be
marked as "implementable".