KEP 2433: Enforce Topology Aware Routing
The existing "Auto" value with the currently defined algorithm
works well for large clusters with services with large endpoints.
However there are situations where that approach doesn't work, namely
low node count multi-region clusters or for service applications that
are not present across all zones within a cluster by design.

This proposal would add an "Always" value to the Topology Aware Hints
annotation to have the controller always sent the Zone hints on the
endpoint slices regardless of node/endpoint count within zones. As long
as an endpoint exists in the zone (or region) then traffic from that
zone should be sent to it.

It would be important for us to ensure we highlight the risks with
overload when we document this feature

/cc @robscott
/sig network