define number of controllers required
There is no need to handle the IPAddresses with finalizers since
the apiserver only deallocates the IP after the Services has been
deleted.

The repair loop on the apiserver will guarantee the relaton between
ServiceCIDR, Services and ClusterIP, there is no need to add a new
controller on the controller manager.