[KEP-2395] Update 2395-removing-in-tree-cloud-providers to beta status (#4190)
* Update metadata for 2395-removing-in-tree-cloud-providers

Signed-off-by: Davanum Srinivas <davanum@gmail.com>

* update KEP-2395 to use most recent template

This adds several sections which were not in the original document. I
have added "TBD" in those sections for now and will update with more
content in followup PRs.

* add beta stage production readiness review approver

* cleanup toc and add release checklist

* update implementation history on kep-2395

* update release signoff checklist for kep-2395

* clean up proposal and design details of kep-2395

As part of the process to adopt the latest template, the implementation
details from the previous `Proposal` section are moved to the new
`Design Details` section. Additionally, the previous `Approach` section
is now in the `Proposal` section with expanded language for the 4th
phase of activity.

* add risks and mitigations to kep-2395

* update test plan for kep-2395

also update the toc

* update testing and release plans after talking with the sig

* update drawbacks section

* update ga target to 1.30

* update toc

* update kep-2395 prr for beta changes

* update kep-2395 with comments from review

this change updates some of the text to match the template and also
addresses concerns brought up on review.

* update prr for kep-2395

* add more testing details to kep-2395

* update ga target for kep-2395

the ga target should be at least 2 releases from the beta because that
is what we say in the kep.

* update prr test info for kep-2395

---------

Signed-off-by: Davanum Srinivas <davanum@gmail.com>
Co-authored-by: Davanum Srinivas <davanum@gmail.com>