DRA: defer to numeric parameters as solution for cluster autoscaling
Technically the support for cluster autoscaling can be defined and implemented
as an extension of the core DRA, without changing the core feature. By
separating out the specification of "numeric parameters" into a separate KEP it
might be easier to make progress on the different aspects because they are
better separated.