docs: add options for KNI controllers to KNI KEP
This adds a section to the KNI KEP for "ongoing considerations"
where we can put ideas and concerns that don't need to be resolved
just yet, but are good to come back to as we progress. It then adds
an idea for using Kubernetes controllers as an alternative to gRPC
APIs for some of the KNI implementation to the ongoing considerations.

Signed-off-by: Shane Utt <shaneutt@linux.com>