DRA: "Structured Parameters" as base, "classic" DRA as extension
We decided to focus on structured parameters as the part of DRA which will get
moved to beta first. "Classic" DRA (PodSchedulingContext, support for control
plane controllers) will remain in alpha. It needs a separate feature gate.

Therefore the description of DRA in the two KEPs gets shifted around such that
the "structured parameters" KEP is a stand-alone, complete description of
everything needed for using DRA like that.

The original KEP is now an extension of that and describes PodSchedulingContext
and the interaction with a DRA driver control plane controller. Immediate
allocation is defined there because it does not make much sense for structured
parameters.

Mostly this was just a matter of copying text around and updating the wording.
Changes that go beyond those editorial changes are:

- title changes

- new diagram for "structured parameters"

- ResourceClass.StructuredParameters in 4381 became
  ResourceClass.ControlPlaneController in 3063 (different default,
  breaking change!)

- new DRAControlPlaneController feature gate for everything in 3063

- removal of the kubelet v1alpha2 gRPC
  interface (https://github.com/kubernetes/kubernetes/pull/124316).

- updated description of ephemeral ResourceClaim namimg (did not match
  the current implementation with generated names)

- added NodeListAndWatchResources kubelet gRPC call