kep-1979: update base design to v1alpha2
Update the COSI KEP's base design to v1alpha2. The primary purpose of
this work is to reshape the doc and spec to fix inconsistencies and
vestigial items, and fix the most glaring issues in the v1alpha1 spec.

COSI maintainers are planning follow-up items to add larger individual
features after this.

Notably, bucketInfo.json has been changed to individual secret fields
with COSI_<KEY>: <VALUE> format, as the JSON blob was flagged as a
problem by several v1alpha1 users.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>