DRA: support abstract attributes
While it isn't really feasible for GPUs, other kinds of hardware might be
described via some standardized attributes. When removing the explicit driver
name from requests and filters it becomes possible for users to ask for an
instance where some standardized attribute is set without knowing in advance
which driver is going to provide it.

The downside is that CEL expressions now must get applied to all instances,
whether they are from the "expected" driver or some other one. It becomes the
responsibility of the CEL expression to filter out unknown instances.