DRA: avoid kubelet API version dependency
By tunneling REST requests from drivers through kubelet to the API server it
becomes possible to update drivers and the control plane to a newer release
than the kubelet because the kubelet doesn't need to encode/decode the
resource.k8s.io API types that are used by the drivers. This simplifies cluster
upgrades.