DRA: avoid kubelet API version dependency
By requiring that drivers on a node connect to the apiserver directly
it becomes possible to update drivers and the control plane to a newer release
than the kubelet because the kubelet doesn't need to encode/decode the
resource.k8s.io API types that are used by the drivers. This simplifies cluster
upgrades.