DRA: optional allocation
Running a pod without some resource may be useful if there is a fallback in the
application code. Allocating more than one may also be useful, in particular
when the upper bound is unlimited. Then it becomes possible to ask for all
instances available on a node in a daemon set.