Add Kubectl Checkpoint KEP
With "Forensic Container Checkpointing" being Beta and discussions
around graduating it to GA, the next step would be kubectl integration
of the container checkpointing functionality.

In addition to the "Forensic Container Checkpointing" use case this KEP
lists multiple use cases how checkpointing containers can be used.

One of the main motivations for this KEP is to make it easier for users
to checkpoint containers, independent of the reason. Having it available
via kubectl reduces the complexity of connecting to the node and
accessing the kubectl checkpoint API endpoint.

Signed-off-by: Adrian Reber <areber@redhat.com>