KEP-2091: Do not pass K8s NetworkPolicy implicit denies to BaselineAdminNetworkPolicy
The previous spec implies that traffic that does not match a rule in any K8s
NetworkPolicy should be passed to the BaselineAdminNetworkPolicy (if present)
and then on to the implementation default behavior (usually Deny).  This would
cause all traffic to be allowed regardless of the NetworkPolicy unless a
BaselineAdminNetworkPolicy is created, which is not an acceptable upgrade
behavior and unlikely to be intended.