KEP-4222: Relax description of decode allocations fuzz test.
The goal of this test was to be able to identify pathological inputs that cause the decoder to
allocate excessive amounts of memory (for example, by trusting the length indicated by the initial
bytes of a string and preallocating a slice of that length). That can be achieved without attempting
to enforce a specific limit on the ratio between input length and bytes allocated, which is subject
to many variables.