KEP-267: kubelet server certificate bootstrap and rotation
This feature has been beta enabled by default since 1.12, there is no
much concern about the stability of the feature, however, it requires an
external component to be used.

Kubernetes must provide "battery included" for all its built-in
features.

Graduate the existing feature to GA by providing a controller in the
cloud-controller-manager that users can opt-in to auto approve the CSR
from the nodes.

Change-Id: I7500f4cb6582fdff423e430518d370bcd08f144a
Signed-off-by: Antonio Ojea <aojea@google.com>