KEP template: enhance suggestions for integration and e2e tests
The template gives the impression that both integration and e2e tests must be
written. Let's be a bit more flexible and document that feature owners can opt
out of that if it makes sense.

Linking to the triage tool is not sufficient. We want an empty result there,
but that could also have been caused by not running the tests at all. A link to
the periodic job which included the tests is more important.

To ensure more consistency, the suggested bullet list gets extended to be more
realistic. Otherwise feature owners have to guess how they are meant to fill in
the placeholders.