KEP template: remove checkboxes
The check *lists* are useful. The check *boxes* are not: presumably ticking
them off is meant to document that authors and reviewers have looked at the
list. But that only works once. In following releases, the box is ticked off
whether the related check list item was checked again or not.

These boxes cause confusion repeatedly because some of them need to be ticked
off before the list item is actually done (PRR review/approval). Others would
have to be updated later in a second PR, which causes extra work for approvers.

In practice, most of the KEPs don't seem to bother with ticking these boxes:

    $ git grep -e '- \[[xX]\]' keps | wc -l
    3611
    $ git grep -e '- \[ \]' keps | wc -l
    3821

That may be legitimate in a few cases (GA), but most cases it just shows that
no-one cared. Let's remove them to simplify the KEP process.