KEP 5295: KYAML
This KEP proposes to add a new output format for kubectl, tentatively
called “KYAML” (as in `kubectl get -o kyaml ...`). This format is a
strict subset (aka “dialect”) of standard YAML, and so should be
parseable by the existing ecosystem. This dialect seeks to emphasize
syntactical choices which avoid many of the most common traps in YAML.
For example, unlike standard YAML output, this dialect is not
whitespace-sensitive, which makes it vastly easier to patch correctly in
things like Helm charts.

This KEP further proposes to make KYAML the standard format for all
project-owned documentation and examples.