Avoid confusion between nil list and empty list
In the credential plugin allowlist, consider nil and empty lists an
error. @liggitt has pointed out that treating them differently is
generally avoided in APIs, as it can confuse users.

In order to make the user's intention more explicit, an additional
field, `credentialPluginPolicy` has been added. Its value must be one of
`EnableAll | DisableAll | Allowlist`. `EnableAll` and `DisableAll` are
self-explanatory and do not require an allowlist. `Allowlist` will use
the allowlist as defined in the KEP.

Note that this change considers the following cases to be errors:
- a policy of `EnableAll` with an allowlist provided
- a policy of `DisableAll` with an allowlist provided
- a policy of `Allowlist` with no allowlist provided (i.e. a `nil`
  allowlist`
- a policy of `Allowlist` with an explicitly empty allowlist provided

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>